+++ src/main/java/pawtropolis/persistence/repository/EntityRepository.java
@@ -3,5 +3,5 @@ package pawtropolis.persistence.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pawtropolis.persistence.dto.entity.EntityDTO;
 
-public interface EntityRepository extends JpaRepository<EntityDTO,Long> {
+public interface EntityRepository extends JpaRepository<EntityDTO,Integer> {
 }
