+++ src/main/java/pawtropolis/persistence/repository/item_related/InventoryRepository.java
@@ -3,5 +3,5 @@ package pawtropolis.persistence.repository.item_related;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pawtropolis.persistence.dto.items.InventoryDTO;
 
-public interface InventoryRepository extends JpaRepository<InventoryDTO,Long> {
+public interface InventoryRepository extends JpaRepository<InventoryDTO,Integer> {
 }
